Use ignoreLocation to search entire text
Since the default `location` for Fuse is 0, the current `threshold`
setting limits matches to the first 30 characters of the text. As a
result, if a table cell has more than 30 characters, and a user copies
the entire text into the search, the filter will return 0 rows (when
users would expect *at least* the row they copied from).

See: https://fusejs.io/api/options.html#ignorelocation

An alternative fix would be to tweak the `location` and `threshold`
settings, or allow consumers to set them, but this is simpler for now.

Signed-off-by: Justin Kulikauskas <justin.kulikauskas@gmail.com>